prepare simpleclient as stand-alone package
The goal is to make simpleclient self-contained. The main sigs.k8s.io/boskos
package may depend on it, but not the other way around.

To achieve this, the common and storage package get moved into simpleclient.
The existing packages, including client, are kept to avoid breaking consumers
of the boskos package. This may be overly cautious, but who knows... One
benefit here is that consumers in the repo do not need to be updated.

Loading secrets from a file remains in client/client.go because it
has too many undesirable dependencies.
