+++ simpleclient/common/common_test.go
